Update index.md (#39)
* Create Salesforce-Integration.md

This is the first draft and needs to be validated by someone who has completed the integration.   The information is accurate, but it might not include all the steps specific to the Salesforce Integration.

* Update Integrations.md

adding link to Salesforce-Integration page

* Update index.md

removing the link to the email submit.digital@gov.bc.ca

* Update About.md

Removing the email contact point and replacing with Teams and Discord

* removing some more mentions of submit box email.

* Update Power-BI-Integration.md

* Add or update example forms with the CHEFS team members as form owners. Created new forms for cases where they were not owned by CHEFS team. (#41)

Co-authored-by: “bhuvan-aot” <“bhuvanesh.p@aot-technologies.com”>

* docs: updated the SoAR compliance log and made some edits

* Update SoAR-and-Compliance.md log for 202-06-27

* Create Salesforce-Integration.md (#38)

* Create Salesforce-Integration.md

This is the first draft and needs to be validated by someone who has completed the integration.   The information is accurate, but it might not include all the steps specific to the Salesforce Integration.

* Update Integrations.md

adding link to Salesforce-Integration page

* adding documentation for Integration of sales force external services

* fix typo.

---------

Co-authored-by: “bhuvan-aot” <“bhuvanesh.p@aot-technologies.com”>

---------

Co-authored-by: “bhuvan-aot” <“bhuvanesh.p@aot-technologies.com”>
Co-authored-by: bhuvan-aot <61068301+bhuvan-aot@users.noreply.github.com>
Co-authored-by: Walter Moar <Walter.Moar@gov.bc.ca>